removed devtools::install_github from vignette
The code in the vignettes is re-evaluated during R CMD check, the prior
version of the vignette with the devtools::install_github call resulted
in the package being installed unconditionally.

If install instructions are to be part of the vignette, the code will
need to be commented out so that it is not automatically evaluated.